+++ src/ocl/convolutionocl_fft.cpp
@@ -397,7 +397,7 @@ static float ExecuteFFTKernel(Handle& handle,
     (void)wDesc; // suppress warning
     (void)fwd;   // suppress warning
 
-    int halfw = static_cast<int>(workSpaceSize) / (2 * 2 * sizeof(float));
+    int halfw = static_cast<int>(workSpaceSize) / (2 * 2 * static_cast<int>(sizeof(float)));
     int in_n, in_c, in_h, in_w;
     std::tie(in_n, in_c, in_h, in_w) = miopen::tien<4>(xDesc.GetLengths());
